% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CAPMRes}
\alias{CAPMRes}
\title{CAPM results}
\usage{
CAPMRes(df, rm, siglevel = 0.05, Newey = TRUE)
}
\arguments{
\item{df}{data.frame With mutual fund returns and market benchmark}

\item{rm}{character Name of the market benchmark}
}
\value{
A data.frame with CAPM results
}
\description{
CAPM results
}

