% Generated by roxygen2 (4.0.2): do not edit by hand
\name{equal_weight}
\alias{equal_weight}
\title{Calculate equal weighted portfolio}
\usage{
equal_weight(df, rm)
}
\arguments{
\item{df}{data.frame With returns to be used in creating an equal-weighted
return}

\item{rm}{A character containing the market return}
}
\value{
data.frame Returns a data.frame with dates and the equal-weighted
  returns
}
\description{
Calculate equal weighted portfolio
}

