% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tex_table}
\alias{tex_table}
\title{Creates a .tex table}
\usage{
tex_table(df, dig, out.file, align = NULL, notes = NULL,
  sig_markers = FALSE)
}
\arguments{
\item{df}{data.frame To be turned into table}

\item{dig}{numeric Number of digits to include in table}

\item{out.file}{Filename for the table. Will be placed in folder called
"Plots" two leves up from wd}

\item{notes}{Character vector to be included as footnote in table. Will accept
LaTeX style formatting.}
}
\description{
Creates a .tex table
}

